Invalidate the cache when the stack changes
Previously the buildpack did not invalidate the cache when the stack
changed, which causes the errors seen in #19. This will be particularly
an issue when Heroku-24 is GAed and Heroku-20 deprecated, since
there will be a number of apps changing stack for the first time after
switching to this buildpack.

Now, the stack version is stored in the cache and if it differs from the
current stack version (or the version file is missing since the cache is
from an older version of this buildpack), the cache will be purged.

This implementation is an improved version of what's implemented
for the generic APT buildpack (I'll be backporting those improvements
over to that buildpack shortly):
https://github.com/heroku/heroku-buildpack-apt/blob/4eb4b35d35d0178e5cef73d6998a26ba2c9bbf17/bin/compile#L42-L86

Longer term we should probably re-evaluate whether we need the cache
at all, since from my testing locally it offered very little benefit, and if
anything causes a number of issues (particularly when several APT-using
buildpacks are set on an app at the same time, given the shared Debian
archives directory and install strategy).

Fixes #19.